requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
numpy>=1.26.0
faiss-cpu>=1.8.0
networkx>=3.3
sentence-transformers>=2.7.0
tree-sitter==0.21.3
tree-sitter-languages>=1.10.2
anthropic>=0.25.0
openai>=1.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
pytest>=8.0.0
httpx>=0.27.0
